fix lzo, misc

+++ src/win32/win_common.h
@@ -85,8 +85,8 @@ enum CriticalSection : __int32
 
 struct FastCriticalSection // sizeof=0x8
 {                                                                             // XREF: .data:g_dvarCritSect/r
-        volatile int readCount;
-        volatile int writeCount;                        // XREF: Dvar_FindMalleableVar_0:loc_7BD2E2/r
+        volatile long readCount;
+        volatile long writeCount;                        // XREF: Dvar_FindMalleableVar_0:loc_7BD2E2/r
 };
 
 void __cdecl Sys_Mkdir(const char *path);
